requirements.txt
amqp==2.5.2
anyjson==0.3.3
appdirs==1.4.0
APScheduler==3.2.0
billiard==3.6.3.0
celery==4.4.0
coreapi==2.0.9
Django==2.1.15
django-appconf==1.0.2
django-codemirror2==0.2
django-colorfield==0.1.15
django-compressor==2.2
django-cors-headers==3.0.2
django-filter==2.0.0
django-guardian==1.4.9
django-imagekit==4.0.1
django-libsass==0.7
django-webpack-loader==0.6.0
djangorestframework==3.9.1
djangorestframework-jwt==1.9.0
drf-nested-routers==0.11.1
funcsigs==1.0.2
futures==3.0.5
geojson==2.3.0
gunicorn==19.7.1
itypes==1.1.0
kombu==4.6.7
libsass==0.19.4
Markdown==2.6.7
olefile==0.44
openapi-codec==1.1.7
packaging==16.8
piexif==1.0.13
pilkit==2.0
Pillow==6.2.2
pip-autoremove==0.9.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
PyJWT==1.5.3
pyodm==1.5.4
pyparsing==2.1.10
pytz==2020.1
rcssmin==1.0.6
redis==3.2.0
requests-toolbelt==0.9.1
requests==2.21.0
rfc3987==1.3.7
rjsmin==1.0.12
simplejson==3.10.0
six==1.11.0
strict-rfc3339==0.7
tzlocal==1.3
uritemplate==3.0.0
vine==1.3.0
webcolors==1.5
-e git://github.com/OpenDroneMap/rio-tiler.git#egg=rio-tiler
rio-color==1.0.0
rio-cogeo==1.1.8
rasterio==1.1.5 ; sys_platform == 'linux' or sys_platform == 'darwin'
https://download.lfd.uci.edu/pythonlibs/s2jqpv5t/rasterio-1.1.3-cp37-cp37m-win_amd64.whl ; sys_platform == "win32"
https://download.lfd.uci.edu/pythonlibs/s2jqpv5t/GDAL-3.0.4-cp37-cp37m-win_amd64.whl ; sys_platform == "win32"
Shapely==1.7.0 ; sys_platform == "win32"
eventlet==0.25.1 ; sys_platform == "win32"
pyopenssl==19.1.0 ; sys_platform == "win32"